requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
easyocr==1.6.2
fonttools==4.38.0
idna==3.4
imageio==2.25.1
importlib-resources==5.12.0
imutils==0.5.4
kiwisolver==1.4.4
matplotlib==3.7.0
networkx==3.0
ninja==1.11.1
numpy==1.24.2
opencv-python-headless==4.5.4.60
packaging==23.0
pandas==1.5.3
pdf2image==1.16.2
Pillow==9.4.0
pyclipper==1.3.0.post4
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scikit-image==0.19.3
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sqlparse==0.4.3
tifffile==2023.2.3
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.14
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zipp==3.14.0
